+++ ETMS.Entity/Database/Source/EtClassRecord.cs
@@ -76,7 +76,7 @@ namespace ETMS.Entity.Database.Source
         /// <summary>
         /// 授课课时
         /// </summary>
-        public int ClassTimes { get; set; }
+        public decimal ClassTimes { get; set; }
 
         /// <summary>
         /// 课消金额
